Flytter util for å merge sammenhengende vilkårperiode til vilkårperiode-package for å unngå kobling til stønadsperioder
