+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/stønadsperiode/StønadsperiodeValidering.kt
@@ -1,9 +1,6 @@
 package no.nav.tilleggsstonader.sak.vilkår.stønadsperiode
 
 import no.nav.tilleggsstonader.kontrakter.felles.Datoperiode
-import no.nav.tilleggsstonader.kontrakter.felles.mergeSammenhengende
-import no.nav.tilleggsstonader.kontrakter.felles.overlapperEllerPåfølgesAv
-import no.nav.tilleggsstonader.sak.felles.domain.FaktiskMålgruppe
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeil
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvis
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvisIkke
@@ -13,8 +10,9 @@ import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.MålgruppeType
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.ResultatVilkårperiode
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeType
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperioder
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.mergeSammenhengendeOppfylteAktiviteter
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.mergeSammenhengendeOppfylteMålgrupper
 
 object StønadsperiodeValidering {
     /**
@@ -98,39 +96,3 @@ object StønadsperiodeValidering {
             }
     }
 }
-
-/**
- *  @return En sortert map kategorisert på periodetype med de oppfylte vilkårsperiodene. Periodene slåes sammen dersom
- *  de er sammenhengende, også selv om de har overlapp.
- */
-inline fun <reified T : VilkårperiodeType> List<Vilkårperiode>.mergeSammenhengendeOppfylte(): Map<T, List<Datoperiode>> =
-    this
-        .filter { it.resultat == ResultatVilkårperiode.OPPFYLT }
-        .groupBy {
-            require(it.type is T) { "${it.type} er ikke av type ${T::class.simpleName}" }
-            it.type
-        }.mapValues {
-            it.value
-                .sorted()
-                .map { Datoperiode(fom = it.fom, tom = it.tom) }
-                .mergeSammenhengende { a, b -> a.overlapperEllerPåfølgesAv(b) }
-        }
-
-fun List<Vilkårperiode>.mergeSammenhengendeOppfylteAktiviteter(): Map<AktivitetType, List<Datoperiode>> =
-    this.mergeSammenhengendeOppfylte<AktivitetType>()
-
-fun List<Vilkårperiode>.mergeSammenhengendeOppfylteMålgrupper(): Map<MålgruppeType, List<Datoperiode>> =
-    this.mergeSammenhengendeOppfylte<MålgruppeType>()
-
-fun List<Vilkårperiode>.mergeSammenhengendeOppfylteFaktiskeMålgrupper(): Map<FaktiskMålgruppe, List<Datoperiode>> =
-    this
-        .filter { it.resultat == ResultatVilkårperiode.OPPFYLT }
-        .groupBy {
-            require(it.type is MålgruppeType) { "${it.type} er ikke av type ${MålgruppeType::class.simpleName}" }
-            it.type.faktiskMålgruppe()
-        }.mapValues {
-            it.value
-                .sorted()
-                .map { Datoperiode(fom = it.fom, tom = it.tom) }
-                .mergeSammenhengende { a, b -> a.overlapperEllerPåfølgesAv(b) }
-        }
